PDeinterlacer: Remove the default QTGMC arguments
The problem is that if someone were to overwrite one of these arguments, it may or may not work as the user may expect. For example if you set the Preset to VeryFast, the user may assume the rest will be the VeryFast preset defaults, not the other defaults shown. This can be very destructive if the user does not know the defaults being used.

So instead of hardcoding my preferred defaults, let the user have full control over the defaults.