PSourcer: Be less restrictive on the file path's type/kind
It's too much bother having to verify what file is coming in, and often times the python built in mimetype check simply isn't vast enough for video extensions.

So to remove a barrier it's best to just get rid of most of the checks instead of having to constantly update a whitelist mapping extensions to file types.