PSourcer: Implement configurable d2v VST VFR matching
The code until now would only frame match VST/VFR content by duplicating progressive frames with the "rff" flag. However that may not be wanted, instead you might want to decimate the interlaced sections to be the same FPS as the progressive sections.

This new code allows you to do exactly that! By default it is `False` meaning it will frame match like it did before this addition. The doc-string explains the possible values for the new parameter.