PD2V: Fix interlaced section enter/exit index offsetting
When entering and exiting an interlaced section, the offset of the last previous progressive section's final index would jump forward in index number quite a bit the next time it enters a progressive section. This causes the two indexes `right-left` calculation to differ for no reason. This is one of the reasons a most-common approach is used. This is now fixed by applying the right-left calculation per-section.

However, random different calculated cycles still appear, so a most-common approach is still needed. Just it's a bit more accurate now, especially for videos with very low rff (but >1 rff).