deps: Remove explicit VapourSynth dependency
This causes issues because the PyPI VapourSynth package is quite heavily out of date and doesn't seem to be the generally intended way to install VapourSynth (anymore maybe?).

Since it's quite outdated and not I guess the "primary" installation method, this causes problems with dependency resolving on newer versions, so best to just remove it.